Backend/requirements.txt
# FastAPI核心框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
SQLAlchemy==2.0.23
aiosqlite==0.19.0
# Supabase 和 PostgreSQL 支持
supabase==2.3.4
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic设置
pydantic-settings==2.1.0

# JWT认证
PyJWT==2.8.0

# 密码加密
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

# 日志管理
loguru==0.7.2

# HTTP请求
requests==2.31.0
httpx>=0.28.1

# 航班搜索库
smart-flights==7.0.0

# 时间处理
python-dateutil==2.8.2

# 环境变量
python-dotenv>=1.0.1

# AI服务
google-generativeai==0.3.2

# 异步支持
asyncio-mqtt==0.13.0

# 开发工具
pytest==7.4.3
pytest-cov==4.1.0

# 缓存
redis==5.0.1

# 其他工具
click==8.1.7
python-multipart==0.0.20
